% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stan_model.R
\name{preload_models}
\alias{preload_models}
\title{preload_models}
\usage{
preload_models()
}
\description{
This R function creates preloaded base models which will make the classification of strong ties faster.
}
\details{
Preload base models for model-based extraction
}
